commit check for saving
