--- src/Footer.js
@@ -1,9 +1,0 @@
-const Footer = () => {
-    return (
-        <div>
-            This is React component
-        </div>
-    );
-}
-
-export default Footer;
