--- src/Header.js
@@ -1,13 +1,0 @@
-const Header = () => {
-    return (
-        <div>
-            <a href='#'> - Home</a>
-            <a href='#'> - News</a>
-            <a href='#'> - Message</a>
-            <a href='#'> - Feed</a>
-        </div>
-    );
-}
-
-
-export default Header;
