+++ src/components/Header.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = () => {
+    return (
+        <header className='header'>
+            <img src="https://img.freepik.com/free-vector/bird-colorful-logo-gradient-vector_343694-1365.jpg?size=338&ext=jpg&ga=GA1.1.2008272138.1724716800&semt=ais_hybrid" alt="" />
+        </header>
+    )
+};
+
+export default Header;
