+++ src/components/Navbar.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Navbar = () => {
+    return (
+        <nav className='nav'>
+            <div><a href="">Profile</a></div>
+            <div><a href="">Messages</a></div>
+            <div><a href="">News</a></div>
+            <div><a href="">Music</a></div>
+            <div><a href="">Settings</a></div>
+        </nav>
+    )
+};
+
+export default Navbar;
