+++ src/components/Profile.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const Profile = () => {
+    return (
+        <div className='content'>
+            <div>
+                <img src="https://t4.ftcdn.net/jpg/06/41/10/41/360_F_641104140_lkHlULzw6I1zwaN4I2I3SN3UNHVmawNW.jpg" alt="" />
+            </div>
+            <div>ava+discription</div>
+            <div>My posts</div>
+            <div>New post</div>
+            <div>
+                <div>post 1</div>
+                <div>post 2</div>
+            </div>
+        </div>
+    )
+};
+
+export default Profile;
