+++ tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-  ]
+, "src/components/Profile.tsx"  ]
 }
